Player: add getMoveSpeed(), change speed factors in Fast, Normal and Slow.

+++ src/unsw/dungeon/Slow.java
@@ -2,10 +2,10 @@ package unsw.dungeon;
 
 public class Slow implements MoveSpeed {
 	// set to 1 for now
-	private int speed = 1;
+	private long speed = 1;
 
 	@Override
-	public int getSpeed() {
+	public long getSpeed() {
 		return this.speed;
 	}
